requirements.txt
fastapi
mysql-connector-python
uvicorn
pydantic